Implement tcp+udp stats, using existing Solaris code
