+++ oshi-core/src/main/java/oshi/driver/unix/NetStat.java
@@ -31,6 +31,8 @@ import java.util.List;
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.software.os.InternetProtocolStats.IPConnection;
 import oshi.software.os.InternetProtocolStats.TcpState;
+import oshi.software.os.InternetProtocolStats.TcpStats;
+import oshi.software.os.InternetProtocolStats.UdpStats;
 import oshi.util.ExecutingCommand;
 import oshi.util.ParseUtil;
 import oshi.util.tuples.Pair;
@@ -40,6 +42,8 @@ import oshi.util.tuples.Pair;
  */
 @ThreadSafe
 public final class NetStat {
+
+    private static final String NETSTAT_S = "netstat -s -";
 
     private NetStat() {
     }
@@ -119,5 +123,138 @@ public final class NetStat {
             }
         }
         return new Pair<>(new byte[0], 0);
+    }
+
+    /**
+     * Queries {@code netstat -s} for the tcp protocol
+     *
+     * @param protocolSwitch
+     *            the netstat command switch character for protocol
+     * @return the stats
+     */
+    public static TcpStats queryTcpStats(Character protocolSwitch) {
+        long connectionsEstablished = 0;
+        long connectionsActive = 0;
+        long connectionsPassive = 0;
+        long connectionFailures = 0;
+        long connectionsReset = 0;
+        long segmentsSent = 0;
+        long segmentsReceived = 0;
+        long segmentsRetransmitted = 0;
+        long inErrors = 0;
+        long outResets = 0;
+        List<String> netstat = ExecutingCommand.runNative(NETSTAT_S + protocolSwitch + " tcp");
+        // append IP
+        netstat.addAll(ExecutingCommand.runNative(NETSTAT_S + protocolSwitch + " ip"));
+        for (String s : netstat) {
+            // Two stats per line. Split the strings by index of "tcp"
+            String[] stats = NetStat.splitOnPrefix(s, "tcp");
+            // Now of form tcpXX = 123
+            for (String stat : stats) {
+                if (stat != null) {
+                    String[] split = stat.split("=");
+                    if (split.length == 2) {
+                        switch (split[0].trim()) {
+                        case "tcpCurrEstab":
+                            connectionsEstablished = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpActiveOpens":
+                            connectionsActive = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpPassiveOpens":
+                            connectionsPassive = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpAttemptFails":
+                            connectionFailures = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpEstabResets":
+                            connectionsReset = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpOutSegs":
+                            segmentsSent = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpInSegs":
+                            segmentsReceived = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpRetransSegs":
+                            segmentsRetransmitted = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "tcpInErr":
+                            // doesn't have tcp in second column
+                            inErrors = ParseUtil.getFirstIntValue(split[1].trim());
+                            break;
+                        case "tcpOutRsts":
+                            outResets = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        default:
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+        return new TcpStats(connectionsEstablished, connectionsActive, connectionsPassive, connectionFailures,
+                connectionsReset, segmentsSent, segmentsReceived, segmentsRetransmitted, inErrors, outResets);
+    }
+
+    /**
+     * Queries {@code netstat -s} for the udp protocol
+     *
+     * @param protocolSwitch
+     *            the netstat command switch character for protocol
+     * @return the stats
+     */
+    public static UdpStats queryUdpStats(Character protocolSwitch) {
+        long datagramsSent = 0;
+        long datagramsReceived = 0;
+        long datagramsNoPort = 0;
+        long datagramsReceivedErrors = 0;
+        List<String> netstat = ExecutingCommand.runNative(NETSTAT_S + protocolSwitch + " tcp");
+        // append IP
+        netstat.addAll(ExecutingCommand.runNative(NETSTAT_S + protocolSwitch + " ip"));
+        for (String s : netstat) {
+            // Two stats per line. Split the strings by index of "udp"
+            String[] stats = NetStat.splitOnPrefix(s, "udp");
+            // Now of form udpXX = 123
+            for (String stat : stats) {
+                if (stat != null) {
+                    String[] split = stat.split("=");
+                    if (split.length == 2) {
+                        switch (split[0].trim()) {
+                        case "udpOutDatagrams":
+                            datagramsSent = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "udpInDatagrams":
+                            datagramsReceived = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "udpNoPorts":
+                            datagramsNoPort = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        case "udpInErrors":
+                            datagramsReceivedErrors = ParseUtil.parseLongOrDefault(split[1].trim(), 0L);
+                            break;
+                        default:
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+        return new UdpStats(datagramsSent, datagramsReceived, datagramsNoPort, datagramsReceivedErrors);
+    }
+
+    private static String[] splitOnPrefix(String s, String prefix) {
+        String[] stats = new String[2];
+        int first = s.indexOf(prefix);
+        if (first >= 0) {
+            int second = s.indexOf(prefix, first + 1);
+            if (second >= 0) {
+                stats[0] = s.substring(first, second).trim();
+                stats[1] = s.substring(second).trim();
+            } else {
+                stats[0] = s.substring(first).trim();
+            }
+        }
+        return stats;
     }
 }
